assignments/Simple AI Agent/requirements.txt
Flask==2.0.1
nltk==3.6.3
gunicorn==20.1.0
cfenv==0.5.3
Jinja2==3.0.1
Werkzeug==2.0.1
openai==1.3.0
requests==2.26.0
python-dotenv==0.19.0